Added policy to controllers methods

+++ app/Http/Controllers/BuslineController.php
@@ -17,6 +17,7 @@ class BuslineController extends Controller
      */
     public function index(): JsonResponse
     {
+        $this->authorize('view');
         $buslines = Busline::all();
 
         return response()->json($buslines);
@@ -27,6 +28,7 @@ class BuslineController extends Controller
      */
     public function store(StoreBusline $request): JsonResponse
     {
+        $this->authorize('manage');
         Busline::create($request->validated());
 
         return response()->json(['message' => __('messages.busline.created')], Response::HTTP_CREATED);
@@ -37,6 +39,8 @@ class BuslineController extends Controller
      */
     public function show(Busline $busline): JsonResponse
     {
+        $this->authorize('view');
+
         return response()->json($busline);
     }
 
@@ -45,6 +49,7 @@ class BuslineController extends Controller
      */
     public function update(UpdateBusline $request, Busline $busline): JsonResponse
     {
+        $this->authorize('manage');
         $busline->update($request->validated());
 
         return response()->json(['message' => __('messages.busline.updated')]);
@@ -55,6 +60,7 @@ class BuslineController extends Controller
      */
     public function destroy(Busline $busline): JsonResponse
     {
+        $this->authorize('manage');
         $busline->delete();
 
         return response()->json(['message' => __('messages.busline.deleted')]);
