+++ app/Http/Controllers/GroupController.php
@@ -17,6 +17,7 @@ class GroupController extends Controller
      */
     public function index(): JsonResponse
     {
+        $this->authorize('view');
         $groups = Group::all();
 
         return response()->json($groups);
@@ -27,6 +28,7 @@ class GroupController extends Controller
      */
     public function store(StoreGroup $request): JsonResponse
     {
+        $this->authorize('manage');
         Group::create($request->validated());
 
         return response()->json(['message' => __('messages.group.created')], Response::HTTP_CREATED);
@@ -37,6 +39,7 @@ class GroupController extends Controller
      */
     public function show(Group $group): JsonResponse
     {
+        $this->authorize('view');
         return response()->json($group);
     }
 
@@ -45,6 +48,7 @@ class GroupController extends Controller
      */
     public function update(UpdateGroup $request, Group $group): JsonResponse
     {
+        $this->authorize('manage');
         $group->update($request->validated());
 
         return response()->json(['message' => __('messages.group.updated')]);
@@ -55,6 +59,7 @@ class GroupController extends Controller
      */
     public function destroy(Group $group): JsonResponse
     {
+        $this->authorize('manage');
         $group->delete();
 
         return response()->json(['message' => __('messages.group.deleted')]);
